Split out Die() and Roller() for complexity
